flybird/requirements.txt
affinity==0.1.0
attrs==17.4.0
Automat==0.6.0
backports-abc==0.5
certifi==2017.11.5
chardet==3.0.4
constantly==15.1.0
DBUtils==1.2
hyperlink==17.3.1
idna==2.6
incremental==17.5.0
itsdangerous==0.24
kafka-python==1.2.5
Markdown==2.6.9
Naked==0.1.31
pycrypto==2.6.1
pymongo==3.6.1
pytz==2017.3
PyYAML==3.12
redis==2.10.6
requests==2.21.0
shellescape==3.4.1
singledispatch==3.4.0.3
six==1.11.0
tornado==4.5.3
Twisted==17.9.0
urllib3==1.22
Werkzeug==0.14.1
zope.interface==4.4.3
ws4py==0.4.2
protobuf==3.5.2.post1
redlock==1.2.0
werkzeug==0.14.1
xlrd==1.1.0
colorlog==3.1.2
aliyun-python-sdk-core==2.8.2
aliyun-python-sdk-ecs==4.8.0
APScheduler==3.5.1
colorlog==3.1.2
oss2==2.6.0
shortuuid==0.5.0
celery==3.1.25
bson